chore: separated test classes into different files
